// src/app/Model/ResourceHistory.ts
export interface ResourceHistory {
    resourceHistoryId: number;
    resourceName: string;
    designation: string;
    resourceCode: string;
    platform: string;
    location: string;
    engagementPlan: string;
    experience: number;
    resourceAllocationDate: Date;
    excelAllocationDate: Date;
    phoneNo: string;
    email: string;
    fileName: string;
  }
  